python_backend/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-cors
gunicorn